feat: ControllerAdvice 바깥에서 ErrorResponse 객체를 생성할 수 있다
 - 테스트 가능하도록 ErrorResponse 정의를 ControllerAdvice 바깥으로
   빼낸다
 - 분리된 ErrorResponse에 정의된 메서드 접근 레벨을 조정해서
   ControllerAdvice에서 호출할 수 있도록 한다
